Move FloatingUI "instantiation" to its own preset hook
