+++ packages/react-sdk/src/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './useRtcStats';
+export * from './useFloatingUIPreset';
