+++ packages/react-sdk/src/hooks/useFloatingUIPreset.ts
@@ -1,0 +1,48 @@
+import { useEffect } from 'react';
+import {
+  offset,
+  autoUpdate,
+  size,
+  useFloating,
+  shift,
+} from '@floating-ui/react';
+import type { UseFloatingProps } from '@floating-ui/react';
+
+export const useFloatingUIPreset = ({
+  placement,
+  strategy,
+}: Pick<UseFloatingProps, 'placement' | 'strategy'>) => {
+  const {
+    refs,
+    x,
+    y,
+    update,
+    elements: { domReference, floating },
+  } = useFloating({
+    placement,
+    strategy,
+    middleware: [
+      offset(10),
+      shift(),
+      size({
+        padding: 10,
+        apply({ availableHeight, elements }) {
+          Object.assign(elements.floating.style, {
+            maxHeight: `${availableHeight}px`,
+          });
+        },
+      }),
+    ],
+  });
+
+  // handle window resizing
+  useEffect(() => {
+    if (!domReference || !floating) return;
+
+    const cleanup = autoUpdate(domReference, floating, update);
+
+    return () => cleanup();
+  }, [domReference, floating, update]);
+
+  return { refs, x, y, domReference, floating, strategy };
+};
